Do not check that explicitly-set engine or previewer is executable
The executable check does not work for engine or preview commands that
contain arguments, like `open -a Preview`. Since the defaults are solely
executables the check can be maintained for them, but for an explicitly
set engine the user will probably have knowledge of whether the
engine/previewer can be executed anyway.

Closes #108